requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
torch>=2.0.0
pyyaml>=6.0
pytest>=7.4.0
lightkurve>=2.4.0
batman-package>=2.4.9
